(DDF-2199) Added an IngestLogger that tracks failed ingest attempts
The new IngestLogger does its best to catch any failed ingest attempts
(more specifically, create requests that die in CatalogFrameworkImpl).
Create requests that fail are loggid in data/log/ingest_error.log and
include the metacards that failed (ID and Title) and a stack trace
of the error that caused the failure.