DDF-76 Fixed a bug where Solr Catalog failed to parse certain queries
Due to a bug in how Solr parses sorted queries, geospatial query operands
must be the first operands in a query expression or else the parser will
error out. This fix makes sure that geospatial query operands are moved to
the front of any operand list before a full query expression gets built.