DDF-186 updated status components
Fixed the mapping of services with metatypes to remove duplicate entries. Added a plugin to make the configuration api somewhat extensible. Updated the source plugin so that it'd be able to work again federated sources and catalog providers. Also added the ability for it to make a best guess on which source goes to which configuration without a configuration pid being present.